small refactor to ttnn-to-emitc pass
