+++ include/ttmlir/Conversion/Passes.td
@@ -7,10 +7,10 @@
 
 include "mlir/Pass/PassBase.td"
 
-def ConvertTTNNToEmitC : Pass<"convert-ttnn-to-emitc", "::mlir::func::FuncOp"> {
-  let summary = "Convert TTNN dialect to EmitC dialect.";
-  let constructor = "createConvertTTNNToEmitCPass()";
-  let dependentDialects = ["mlir::emitc::EmitCDialect", "mlir::tt::ttnn::TTNNDialect"];
+def ConvertTosaToTTIR : Pass<"convert-tosa-to-ttir", "::mlir::ModuleOp"> {
+  let summary = "Convert TOSA dialect to TTIR dialect.";
+  let constructor = "createConvertTosaToTTIRPass()";
+  let dependentDialects = ["mlir::tt::ttir::TTIRDialect"];
 }
 
 def ConvertTTIRToTTNN: Pass<"convert-ttir-to-ttnn", "::mlir::ModuleOp"> {
@@ -19,10 +19,10 @@ def ConvertTTIRToTTNN: Pass<"convert-ttir-to-ttnn", "::mlir::ModuleOp"> {
   let dependentDialects = ["mlir::tt::ttir::TTIRDialect", "mlir::tt::ttnn::TTNNDialect"];
 }
 
-def ConvertTosaToTTIR : Pass<"convert-tosa-to-ttir", "::mlir::ModuleOp"> {
-  let summary = "Convert TOSA dialect to TTIR dialect.";
-  let constructor = "createConvertTosaToTTIRPass()";
-  let dependentDialects = ["mlir::tt::ttir::TTIRDialect"];
+def ConvertTTNNToEmitC : Pass<"convert-ttnn-to-emitc", "::mlir::ModuleOp"> {
+  let summary = "Convert TTNN dialect to EmitC dialect.";
+  let constructor = "createConvertTTNNToEmitCPass()";
+  let dependentDialects = ["mlir::emitc::EmitCDialect", "mlir::tt::ttnn::TTNNDialect"];
 }
 
 #endif // TTMLIR_CONVERSION_PASSES
