+++ lib/Conversion/TTNNToEmitC/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_mlir_library(TTMLIRTTNNToEmitC
   TTNNToEmitC.cpp
+  TTNNToEmitCPass.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${PROJECT_SOURCE_DIR}/include/ttmlir/Conversion/TTNNToEmitC
